+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 ## v1.2.2 (no planned release date)
 
+#### Changes
+
+- Electron has been upgraded to v9.1.2
+
 #### Bugs fixed
 
 - Configuring the git signature on initial setup has been fixed (#2)
 - Main navbar should no longer scroll out of view (#3)
+- On MacOS, all hotkeys now use the ⌘ key instead of Ctrl (#5)
 
 ## v1.2.1 (2020-05-10)